% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparesnapshot.R
\name{CompareSnapshot}
\alias{CompareSnapshot}
\title{CompareSnapshot}
\usage{
CompareSnapshot(widget, diff.file, accepted.file, delay = 0.2,
  threshold = 0.1, strict = FALSE)
}
\arguments{
\item{widget}{The htmlwidget to display.}

\item{diff.file}{Character; the name of the snapshot file taken}

\item{accepted.file}{Character; the name of the snapshot to be compared against.
If this file does not exist, it will be created from a snapshot of the widget.}

\item{delay}{Time to wait before taking screenshot, in seconds. Sometimes a longer delay is needed for all assets to display properly.}

\item{threshold}{similarity parameter for comparing screenshots.}

\item{strict}{Whether to give warning if \code{accepted.file} does not exist.}
}
\value{
TRUE if the accepted snapshot does not exist or the widget 
  looks the same as the accepted snapshot
}
\description{
Takes snapshot of a widget and compares it against another snapshot.
}
