% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testwidget.R
\name{TestWidget}
\alias{TestWidget}
\title{TestWidget}
\usage{
TestWidget(widget, name, delay = 0.2, threshold = 0.1,
  diff.path = "snapshots/diff", accepted.path = "snapshots/accepted")
}
\arguments{
\item{widget}{The htmlwidget to display.}

\item{delay}{Time to wait before taking screenshot, in seconds. Sometimes a longer delay is needed for all assets to display properly.}

\item{threshold}{Similarity parameter for comparing screenshots.}

\item{diff.path}{Directory to place diff snapshot.}

\item{accepted.path}{Directory to expect accepted snapshots to be.
If no snapshots are found here, new snapshots will be taken of the
widget and these will be considered the new reference.}
}
\value{
Return \code{FALSE} only if the snapshot of the widget differs from
 the accepted snapshot (i.e. returns \code{TRUE} when the accepted snapshot
 does not exist). The \code{diff.file} is only created if \code{FALSE} is 
 returned otherwise the snapshot will be saved as \code{accepted.file}.
}
\description{
A wrapper for CompareSnapshot for convenience in testthat.
Compares a widget against a reference snapshot, creating snapshot if
 it differs from the reference or the reference does not exist.
}
\seealso{
\code{link{CompareSnapshot}}
}
