% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparesnapshot.R
\name{CompareSnapshot}
\alias{CompareSnapshot}
\title{CompareSnapshot}
\usage{
CompareSnapshot(widget, diff.file, accepted.file, delay = 0.2,
  threshold = 0.1, strict = FALSE)
}
\arguments{
\item{widget}{The htmlwidget to display.}

\item{diff.file}{Character; the name of the snapshot file taken. Only saved if there is a difference.}

\item{accepted.file}{Character; the name of the snapshot to be compared against.
If this file does not exist, it will be created from a snapshot of the widget.}

\item{delay}{Time to wait before taking screenshot, in seconds. Sometimes a longer delay is needed for all assets to display properly.}

\item{threshold}{Similarity parameter for comparing screenshots.}

\item{strict}{Whether to give warning if \code{accepted.file} does not exist.}
}
\value{
Return \code{FALSE} only if the snapshot of the widget differs from
 the accepted snapshot (i.e. returns \code{TRUE} when the accepted snapshot
 does not exist). The \code{diff.file} is only created if \code{FALSE} is 
 returned otherwise the snapshot will be saved as \code{accepted.file}.
}
\description{
Takes snapshot of a widget and compares it against another snapshot.
}
