% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createsnapshot.R
\name{CreateSnapshot}
\alias{CreateSnapshot}
\title{CreateSnapshot}
\usage{
CreateSnapshot(
  widget,
  filename,
  delay = 0.2,
  width = 992,
  height = 744,
  mouse.hover = TRUE,
  mouse.click = FALSE,
  mouse.doubleclick = FALSE,
  mouse.xpos = 0.5,
  mouse.ypos = 0.5,
  debug = FALSE
)
}
\arguments{
\item{widget}{The htmlwidget to display.}

\item{filename}{Name of output file (png format).}

\item{delay}{Time to wait before taking screenshot, in seconds. Sometimes a longer delay is needed for all assets to display properly.}

\item{width}{Width of the snapshot viewport in pixels.}

\item{height}{Height of the snapshot viewport in pixels.}

\item{mouse.hover}{Logical indicating whether or not to simulate mouse moving.
Note that the cursor itself is not visible, only the effects of clicking.}

\item{mouse.click}{Logical indicating whether to simulate mouse click.
Overrides \code{mouse.click}.}

\item{mouse.doubleclick}{Logical indicating whether to simulate doubleclicking.
Overrides \code{mouse.hover} and \code{mouse.click}.}

\item{mouse.xpos}{Horizontal position of the mouse ranging from 0 (left) to 1 (right).}

\item{mouse.ypos}{Vertical position of the mouse ranging from 0 (top) to 1 (bottom).}

\item{debug}{Print debug message from snapshotting.}
}
\description{
Converts a htmlwidget to png image file
}
\details{
Works with plotly and rhtmlLabeledScatter. Errors with rhtmlPictograph.
}
